Exploring example "contract"
Am exploring an example contract (in test.dfy), where I am able to write
a postcondition on the returndata.  That is (in principle) quite neat!
However, there be dragons here.  At this time, using the MSTORE bytecode
causes Dafny to go into a spin.  Using MSTORE8, however, is fine.

The next action here is probably to explore what we can do about MSTORE.
Also, I think the API is not setup well for a gradual translation
approach.