Rewrite Memory.write_u256()
This method is used by MSTORE, and was suspected of being the reason
for the very poor MSTORE performance.  The method original used
bitvectors to do the necessary shifting.  However, this turns out to a
poor choice and straightforward arithmetic based on division and
remainder is far more efficient (in Z3).

This change now means the sample contract in test.dfy passes
verification, though there are still some unresolved questions.  For
example, why we cannot prove the size of the resulting returndata is
0x20.