Refactor Memory.slice #2
This refactors Memory.slice to use a subdivision method, rather than an
iterative style.  The theory was that this would reduce work on the
theorem prover.  In fact, that does seem to be the case as I can now
verify my test contract again.

Despite this interesting success, I still think it might be nice to
try reworking Memory to use a seq<u8>.  My feeling is that this will
still offer the best overall performance.