Support SLT, SGT
This adds support for the SLT and SGT bytecodes, along with a minimal
number of tests.  To do this required adding a method to convert from a
"word" to a signed 256 integer.

In addition, this fixes a bug related to the ordering of operands for
arithmetic operations (ADD/SUB/MUL/DIV).  The ordering was the wrong way
around!