Fix #17
This adds imeplementations for AND/OR/XOR and NOT bytecodes.  These are
implemented by converting to bitvectors.  However, there appear to be
some problems with OR and XOR for reasons unknown.  Specifically, they
cause verification to timeout.  So, both implementations are marked
{:verify false} and a separate issue create to resolve this (see #23).

In addition, a small number of tests have been added to verify correct
operation.