Fix #79
This pushes through a fix for running the state tests.  Specifically, it
now correctly handles the parameterisation found in these tests, where
this exploits "indexes" to determine concrete values for data, gas and
value.

This also includes a number of other fixes to get the tests in stExample
to pass.  At this stage, there is one failure (which arises due to
    missing support for CALLVALUE) and some which are ignored (due to
      lack of machinery for gas calculations).