Support /runAllTests in build.gradle #94
This adds support for using Dafny's runtime testing facility to the
gradle build.  Specifically, it will perform runtime tests on tests
found in the src/test/dafny directory.

Unfortunately, Dafny does not return a non-zero exit code on test
failure.  That is both surprising, and annoying as it means gradle
doesn't detect the test failure.