Add support for PUSHXX bytecodes #117 #21
This adds support for (in principle) the full variety of PUSH bytecodes.
However, there are some dragons.  In particular, the map used within
EvmBerlin does not verify with the full set of bytecodes.  I think we
are hitting some kind of EVM limit here, though it is hard to say.