Call bytecode partially implemented and working.
At this stage, the call bytecode is working but there are certain
features which are missing.  For example, it does not at this time
handle calldata.  However, it does actually handle returndata at least.

Likewise, the code which actually manages the contract call doesn't yet
do enough checks.  For example, that the caller has sufficient balance
for the value being transferred.

In addition, the worldstate is not updated correctly by the execution
environment, and this would need to be fixed.