Support WorldState in GeneralStateTests #180
This adds support for converting the worldstate from the given reference
test instance into the form required for DafnyEvm.  This is necessary
for any tests which involve external calls.

This also updates the version of EvmTools to 0.3.6 as a couple of minor
changes were required there (basically for accessing the worldstate
    data).