Partial support for CREATE #253
In essence, the bytecode appears to work.  However, the problem is that
it doesn't calculate the new contract address according to the official
formula (it uses a much simpler method).  Unfortunately, that means the
state tests will fail because, of course, the contract address is loaded
on the stack upon return (hence all the traces differ at this point).