Fix gas calculation for Keccak256
This also updates the codebase for traces which contain `gas`
information, such that I can start debugging it.  Also included is a
sketch implementation for Blockhash.  The trace comparison is updated so
that traces fail if an exception occurrs on both sides, but the error
code differs.  Finally, have refactored the gas calculations for
expansion such that various bytecodes (log, returndatacopy, etc) are
accounted for.