Implement external SHA3 call.
This puts in place the plumbing necessary to make it work on the Java
side.  However, its fair to say that Dafny does not make this easy and
I'm unclear what the "offcial" way to do this is (if indeed there is
one).

This reuses an existing implementation of Keccak on the Java side, taken
from web3j.

This also fixes some bugs related to the gas calculated and supplied to
a CALL.  However, there is work still to do there.