Fix #303
This now correctly handles contract creation at the transaction level.
In such case, the "to" field is missing from the transaction, and we are
now correctly executing a contract creation.  Previously we were
executing a contract call, providing the calldata as the code.  That
kind of worked, but the address calculation for the recipient was wrong.
Hence, the need for a proper contract creation process which calculates
the correct contract address.