Fix #358
This puts through a fix for the `CALLDEPTH_EXCEEDED` exception.  The
problem is that depth actually starts from zero, but Geth counts from 1.
I managed to confirm this by comparing with Besu and also the
execution-specs.  Besu follows Geth (and presumably EIP3155), whilst the
execution specs after different.

The issue was that Geth allows the depth to get to 1025, which should be
invalid according to the yellow paper.  However, the yellow paper must
presumably be assuming that we count from zero.  Fun times :)