Add gas checks for JUnit tests #185
This updates every test to include a gas check.  That is, a check on
exactly how much is used versus what we are expecting.  At this stage, I
have essentially just implemented this as a regression test rather than
trying to calculate the expected cost of each test (though I did
calculate some).