Minimise occurence of MEMORY_OVERFLOW #378
The `MEMORY_OVERFLOW` is a ficticious exception create to appease Dafny
in certain circumstances.  In particular, they Yellow Paper places no
restrictions on the bounds of memory.  And yet, there are implied
limits.  In particular, the `MSIZE` bytecode loads the length of memory
onto the stack as a `u256`.  But, the yellow paper says nothing about
what should happen in the event of an overflow.