Fix #435
This puts through a fix for handling an outermost CREATE failure.  In
particular, an outermost CREATE cannot attempt to create a contract
which is too large.

The fix for this has slightly altered the way that trace data is
generated from the DafnyEVM.  Overall, I still think there is work to do
here in figuring out the right mechanism.