Complete Blake2bf precompiled contract #340
This completes the precompiled contract.  The issue appears to be that
the sample code provided in EIP152 actually differs from RFC7693.  Its
not clear to me why this is, but for some reason they chose to tweak the
algorithm.

This also fixes some bugs in the encoding/decoding functions which are
now located in the file dafnyevm.util.Bytes.