Simplistic Proof Generator
This implements a very simplistic and naive proof generator as a
proof-of-concept to help assess whether this approach is viable.

The proof generator has now been tested on a small contract originally
written in Solidity and compiled down to Yul.  There are some
outstanding issues related to determining the right number of require
operands and capacity for the block methods.  And also the JUMPDEST
issue is getting in the way a bit.

However, in principle, we can consider the test to be a success.