Update to Dafny 3.10.0 #496
This updates the default Dafny version for building the Dafny EVM to be
3.10.0.  In particular, that required fixing some compatibility issues
introduced (see below).  The Github Action is also updated accordingly.

The compatability problem appears to be a change in the way that Dafny
compiles `datatype` instances into Java.  For example, previously
the Dafny type `Storage.T` was represented using a Java class `T` which
wrapped a `DafnyMap`.  Now (for whatever reason), `Storage.T` is
represented directly as a `DafnyMap`.  Presumably this is only happens
when the `datatype` has only a single case (which `Storage.T` does).