Refactor OKState => ExecutingState #511
This puts in place a number of refactorings which aim to improve the
readability of the code base.  In particular, related to the machine
states.  For example, the notion of an "executing" state vs a
"terminated" state is clarified.

Furthermore, the notion of a continuation is factored into a
"continuing" state which identifies the cases.