Rework DafnyEVM API
This reworks the API so that call and create and unified into a single
method "execute" which accepts a transaction.  In fact, at this time, it
accepts only legacy transactions.

This adds support for the intrinsic gas calculation, and this should
resolve #328.

This finalises the updated transaction execution function for the
DafnyEvm, and likewise resolves issues around transaction failures (e.g.
out of funds, not enough gas, etc)