Preliminary support for Dafny 4 CLI
This adds preliminary (and currently inactive) support in the
`build.gradle` file for building the Java backend using the new Dafny
Command-Line Interface.  Unfortunately, this does not work at the
present time and there is no clear solution.  Two issues have been
raised on the Dafny repository (3611 and 3612).

In addition, have moved the `Array<T>` type into a separate file
(`util/arrays.dfy`) so it can be included in `precompiled.dfy` and
`extern.dfy`.  This doesn't really change much, but offers a minor
improvement in comprehension.