Remove bodies from `{:extern}` from precompiles
This presents a solution (on the Java side) which allows us to remove
the bodies of precompile functions marked `{:extern}`.  The solution
requires a bit of Java hackery, but it works.  In essence, we have a
default implementation which can be overriden.  The default
implementation is used for compiling Dafny, whilst the real
implementation is linked dynamically when the DafnyEvm class is
instantiatied.