Update trace tests
This updates the trace test with the new format used in EvmTools
(v0.3.26).  This format is now recursive and more accurately reflects
the nature of execution within the EVM.  This does expose some
limitations in the way Geth reports trace data, but for the most part it
works pretty well.

A key improvement will be better debugging around contract calls.
However, at this stage, to get that additional information requires
manual tweaks to override code which is attempting to mirror Geth.