Randomisation and log summary
The gradle build now supports randomisation which can be specified via a
parameter (e.g. `gradle build -Prandomize=5`).  This means it will
verify multiple times with different seeds values in an effort to help
identify unstable verification checks.  I have added this by default to
the github action now as well.

This also improves the output from the `DafnyLogSummariser` in the case of
multiple verification runs (i.e. after `/randomSeeIterations:x` is
used).  In particular, it now reports the "Coefficient of Variation".