Implement `MOD_EXP` precompile contract
This implements the `MOD_EXP` precompile contract using the new
`Int.ModPow()` function in Dafny.  This required implementing conversion
methods from a `nat` to the corresponding bytes, and vice-versa.  A
further issue here is that `Int.FromBytes()` is not tail recursive and,
hence, overflows the stack for large sequences.  To address this, have
implemented `Int.FromBytes2()` which takes a divide-and-conquer style
approach.  This reduces the maximum stack height required.  However,
establishing that `FromBytes(x) == FromBytes2(x)` is challenging and I
have left an unproven lemma for this.