Update Dafny / Java code for Dafny 4.2.0
This makes a small number of adjustments to the Dafny code (e.g.
removing unnecessary semi-colons) and the Java code (e.g. Dafny naming
has changed from `EvmState_Compile` to just `EvmState`).  This adds the
DafnyRuntime as a Maven dependency, since it is now available.

At this stage, is not verifying however due presumably to the known
issues with Z3.