Fixes for Go backend
Seems like changes made to the Go backend have introduced a range of
name clashes (as they did for the Java backend).  This partially fixes
that as follows:

(1) Module `Bytes` renamed to `ByteUtils` to avoid clash with `bytes`.
(2) GNUMakefile `dafny_test_global` target commented out, as there
remains a clash on this side which I did not have the patience to
diagnose.
(3) Added simple mechanism for choosing Z3 4.8.5 when the `DAFNY_HOME`
environment variable is set.

At some point, it also probably makes sense to update the makefile for
the Dafny 4 CLI style.