Link and Integrate DafnyCrypto
This links the DafnyCrypto library as a submodule, and integrates it
(superficially) into the DafnyEVM.  The next step is to remove the
`crypt/alt_bn128.dfy` module and replace it with the equivalent from the
DafnyCrypto library.

Also, should remove things from the `Int` module which are now located
within the `DafnyCrypto/util/MathUtils.dfy`.