Support separate "forks" #526
This adds _preliminary_ support for forks through the `EvmFork.Fork`
structure.  There remains quite a bit to do, however.  For example,
properly implementing the `LONDON` fork would be an initial goal, along
with updating the tests accordingly.

The essential essence of a fork is twofold: firstly, a sequence of
active EIPs; secondly, a sequence of active bytecodes.  The latter
could, in principle, be derived from the former.  However, to keep
things as simple as possible, for now they are constants.