Enable `London` fork in `GeneralStateTests`
This enables tests specifically for the `London` fork.  At this stage,
around 40 are failing for a variety of (undiagnosed) reasons.
Furthermore, its unclear whether the DafnyEVM is actually initialised
with the appropriate fork.

This also selects the appropriate `EvmFork.Fork` Dafny construct based
upon the fork string used within a given `StateTest`.  It also
adds the `baseFee` as a field of `BlockInfo` which is populated from the
state test info.