Fix Java Build issues in Dafny `4.4.0`
This puts through a fix for build issues experienced with Dafny `4.4.0`.
The problem stems from a breaking change in the way that Dafny generates
classes from `datatype` definitions in Dafny source files.
Specifically, in this case, the `Option` datatype which is defined in
`option.dfy` as part of the DafnyCrypto library.

The fix for this is somewhat hacky, in that we use `null` instead of
`Option_None`.  However, overall, it seems quite reasonable in this
case.