Partially implement EIP-3860
This puts through a partial implementation of EIP-3860 "Limit and Meter
Initcode".  Specifically, it does the "meter initcode" bit but not yet
the "limit initcode" bit.

This also now marks all tests failing because of missing Shanghai
features as ignored in `GeneralStateTests`.  Mostly seems like on
progress to get Shanghai up and running.